fixing types and adding tests
